renamed loginController.js to gitHubController.js
